### 2024-07-19 #6996 @Nikoleta-v3
> 👋 @Nikoleta-v3 - Here's a second submission that I wonder if you could edit? If you can take this one or the other, it would be great!

@danielskatz 👋🏻 I'm happy to review this submission 😄 Unfortunately, I can't do both at the moment.
