### 2024-07-20 #6996 @Nikoleta-v3
Thank you for your submission and for suggesting reviewers @rwestaway. I have successfully managed to install the package and run a few examples. The package is very well documented, and the readme is very clear and well organized 😄

I am going to contact some of the authors of recent publications first (see https://github.com/openjournals/joss-reviews/issues/6996#issuecomment-2230669081) to see if they can review the paper. In the meantime, could you please address some of my comments and questions?

### Paper

There are some minor typos in the paper:

- Line 27: "behaves"
- Line 35: "postcode"
- Line 72: "by the UK."

Also in the paper, you have two references for `inlabru`. Their Methods in Ecology and Evolution paper and the GitHub repository. Is there a reason for that?

### Tests

I see that you have some tests. Could you include some information on how to run the test suite? I really like your contribution guidelines; they are short and to the point. Maybe adding the test information there? or would be a bit much?

Whatever you decide, it would just be useful to know how to run the tests locally since the tests run when people open pull requests.
