### 2024-07-20 #6996 @Nikoleta-v3
👋🏻 @jhollist @PieterjanRobbe @wcjochem @thomaspinder would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "4DModeller: a spatio-temporal modelling package". You can find more information at the top of this Github issue 🆙
