### 2024-07-22 #6996 @rwestaway
> Thank you for your submission and for suggesting reviewers @rwestaway. I have successfully managed to install the package and run a few examples. The package is very well documented, and the readme is very clear and well organized 😄
>
> I am going to contact some of the authors of recent publications as well (see [#6996 (comment)](https://github.com/openjournals/joss-reviews/issues/6996#issuecomment-2230669081)) to see if they can review the paper. In the meantime, could you please address some of my comments and questions?
>
> ### Paper
> There are some minor typos in the paper:
>
> * Line 27: "behaves"
> * Line 35: "postcode"
> * Line 72: "by the UK."
>
> Also in the paper, you have two references for `inlabru`. Their Methods in Ecology and Evolution paper and the GitHub repository. Is there a reason for that?
>
> ### Tests
> I see that you have some tests. Could you include some information on how to run the test suite? I really like your contribution guidelines; they are short and to the point. Maybe adding the test information there? or would be a bit much?
>
> Whatever you decide, it would just be useful to know how to run the tests locally since the tests run when people open pull requests.

Thank you @Nikoleta-v3 for your prompt feedback. Should we make changes to the paper.md file as we go along or is that now 'frozen' until all reviewers have responded? We haven't submitted to JOSS before so are unsure of the process...

### Paper

Typos will be corrected.

The distinction is between (1) the applications/tutorials of inlabru which are only available via the GitHub repository (cited as Lindgren et al., 2024) and (2) the underlying statistics/method for which the package authors provide two references at https://inlabru-org.github.io/inlabru/authors.html#citation (Bachl et al., 2019 and Yuan et al., 2017). Does that make sense or does it need further clarification in the text?

### Tests

I will have to consult the project team on this one.
