### 2024-07-22 #6996 @mnky9800n
> **Software report:**
>
> ```
> github.com/AlDanial/cloc v 1.90  T=0.08 s (746.5 files/s, 499974.9 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> CSV                              4              0              0          28871
> Rmd                             17           1024           1364           1959
> R                               21            289            532           1848
> TeX                              2            110              0           1077
> XML                              1              0             12            691
> Markdown                         8            185              0            645
> YAML                             4             11             12            210
> JSON                             1              0              0              8
> -------------------------------------------------------------------------------
> SUM:                            58           1619           1920          35309
> -------------------------------------------------------------------------------
> ```
>
> **Commit count by author**:
>
> ```
>    407	Gareth Jones
>     41	Anrijs K. Abele
>     40	XueqingYin
>     18	mnky9800n
>     12	John Aiken
>      7	desireetreichler
>      5	Anrijs Abele
>      4	Tian Li
>      3	Désirée Treichler
>      2	Kristoffer Aalstad
>      1	Alexander Minakov
>      1	El
>      1	Jonathan Bamber
>      1	Julie Røste
>      1	mmazzolini
> ```

FYI @mnky9800n is John Aiken and that is me. so don't know why it counts me separate but it s not a big deal i guess.
