### 2024-07-23 #6996 @Nikoleta-v3
> Thank you @Nikoleta-v3 for your prompt feedback. Should we make changes to the paper.md file as we go along or is that now 'frozen' until all reviewers have responded? We haven't submitted to JOSS before so are unsure of the process...

Either is fine! If you push to the JOSS branch, everything will be updated. To be honest, I used to wait until I found reviewers to give any feedback on the paper, but since finding reviewers can be quite challenging at times, I started giving feedback beforehand so there is not too much time wasted. This is new territory for me too, haha.

I think you are fine to make the changes, at least the paper changes, and push so the reviewers will have the latest version with the corrections.

> ### Paper
> Typos will be corrected.
>
> The distinction is between (1) the applications/tutorials of inlabru which are only available via the GitHub repository (cited as Lindgren et al., 2024) and (2) the underlying statistics/method for which the package authors provide two references at https://inlabru-org.github.io/inlabru/authors.html#citation (Bachl et al., 2019 and Yuan et al., 2017). Does that make sense or does it need further clarification in the text?

Thank you, I understand; I believe it's fine as it is!

> ### Tests
> I will have to consult the project team on this one.

Great; thank you.
