+++ man/paper.md
@@ -46,9 +46,9 @@ bibliography: paper.bib
 
 4DModeller (`fdmr`) is a spatio-temporal modelling package capable of solving a wide range of large-scale space-time (i.e. four-dimensional) problems [@Yin:2023]. It is built around the inlabru framework which is a suite of R codes for fast efficient Bayesian inference [@Yuan:2017; @Bachl:2019]. The `fdmr` package expands the inlabru framework to include specific applications of latent variable modelling for 4-D geophysical problems (e.g. ocean heat content, the Earth’s magnetic field, and global sea-level rise). `fdmr` also includes shiny apps that provide tools for data visualization, finite element mesh building and Bayesian hierarchical modelling based on an R package for Bayesian inference, inlabru, along with model evaluation and assessment. These shiny apps are designed to make the complex INLA framework [@Rue:2009] and associated concepts accessible to a wider scientific community, including users who have little to no previous experience using R. The tools are designed with new users in mind by leveraging their expertise with their data sets while minimizing the need to develop extensive code in R [@Aiken:2018; @Vygotsky:1978]. They allow users to interact with their data first using the intuitive knowledge of the modelling process (input data, create mesh, calculate statistical model), then auto-generating code that the users can build on.
 
-This is extended through the Tutorial Driven Software Development practice [@Woods:2022]. This approach is designed to integrate subject matter experts into the code development cycle. It involves the identification of representative and instructive use cases, followed by tutorials that describe how the package could be used to solve them, and then finally code written and tested so that it behaved as described in the tutorials [@Woods:2022]. `fdmr` users have access to a set of domain-specific tutorials as vignettes in R Markdown notebooks; tutorials which are being added to as the user community grows.
+This is extended through the Tutorial Driven Software Development practice [@Woods:2022]. This approach is designed to integrate subject matter experts into the code development cycle. It involves the identification of representative and instructive use cases, followed by tutorials that describe how the package could be used to solve them, and then finally code written and tested so that it behaves as described in the tutorials [@Woods:2022]. `fdmr` users have access to a set of domain-specific tutorials as vignettes in R Markdown notebooks; tutorials which are being added to as the user community grows.
 
-The current development of the `fdmr` package supports a wide range of spatially heterogeneous and areal data, including in-situ point observations and satellite data. Examples of the former are ground station air pollution observations, rain gauge data, ocean buoy measurements, or GPS ground displacements. For areal data, the domain mesh is fixed (of regular or irregular shape) and partitioned into areal units (e.g. triangles) with well-defined boundaries. Examples of areal data are attributes collected by post code, satellite imagery, spatially gridded products such as climate re-analysis or land use classification.
+The current development of the `fdmr` package supports a wide range of spatially heterogeneous and areal data, including in-situ point observations and satellite data. Examples of the former are ground station air pollution observations, rain gauge data, ocean buoy measurements, or GPS ground displacements. For areal data, the domain mesh is fixed (of regular or irregular shape) and partitioned into areal units (e.g. triangles) with well-defined boundaries. Examples of areal data are attributes collected by postcode, satellite imagery, spatially gridded products such as climate re-analysis or land use classification.
 
 Future package development efforts will focus on expanding its capabilities and broadening its applicability. Moreover, our team actively seeks interdisciplinary collaborations to further expand the modelling framework and tailor it to the specific needs of diverse disciplines.
 
