### 2024-07-24 7090416
Addition of new reference (Lindgren et al., 2024 arXiv)
