+++ man/paper.bib
@@ -40,10 +40,20 @@
   urldate = {2024-05-17}
 }
 
-@Misc{Lindgren:2024,
-  author = {{Lindgren}, F. and {Bachl}, F.E. and {Borchers}, D.L. and {Simpson}, D. and {Scott-Howard}, L. and {Seaton}, A. and {Sue}, M.H. and {Roudier}, P. and {Meehan}, T. and {Peddinenikalva}, N. and {Perepolkin}, D.},
+@Misc{Lindgren:2024a,
+      author={{Lindgren}, F. and {Bachl}, F.E. and {Illian}, J. and {Suen}, M.H. and {Rue}, H. and {Seaton}, A.E.},
+      title={inlabru: software for fitting latent Gaussian models with non-linear predictors},
+      year={2024a},
+      eprint={2407.00791},
+      archivePrefix={arXiv},
+      primaryClass={stat.ME},
+      url={https://arxiv.org/abs/2407.00791}, 
+}
+
+@Misc{Lindgren:2024b,
+  author = {{Lindgren}, F. and {Bachl}, F.E. and {Borchers}, D.L. and {Simpson}, D. and {Scott-Howard}, L. and {Seaton}, A. and {Suen}, M.H. and {Roudier}, P. and {Meehan}, T. and {Peddinenikalva}, N. and {Perepolkin}, D.},
   title = {inlabru},
-  year = {2024},
+  year = {2024b},
   month = {},
   journal = {GitHub repository},
   url = {https://inlabru-org.github.io/inlabru},
