### 2024-07-26 #6996 @rwestaway
> > ### Tests
> > I will have to consult the project team on this one.
>
> Great; thank you.

@Nikoleta-v3 - To respond to your query, the testing suite is a relic of some old code so thank you for flagging that! At the time we had some ideas for testing meshes, and other topics, but as it stands we haven't really built it out or developed it further. However it is something planned for future work.
