### 2024-07-26 35f148a
Revert "Addition of new reference (Lindgren et al., 2024 arXiv)"
