+++ man/paper.bib
@@ -1,156 +1,0 @@
-@article{Aiken:2018,
-  author={{Aiken}, J.M. and {Aiken}, C. and {Cotton}, F.},
-  title = "{A Python Library for Teaching Computation to Seismology Students}",
-  journal = {Seismological Research Letters},
-  volume = {89},
-  number = {3},
-  pages = {1165-1171},
-  year = {2018},
-  month = {03},
-  issn = {0895-0695},
-  doi = {10.1785/0220170246},
-  url = {https://doi.org/10.1785/0220170246},
-  eprint = {https://pubs.geoscienceworld.org/ssa/srl/article-pdf/89/3/1165/4127200/srl-2017246.1.pdf}
-}
-
-@Article{Bachl:2019,
-  author = {{Bachl}, F.E. and {Lindgren}, F. and {Borchers}, D.L. and {Illian}, J.B},
-  title = {{inlabru}: an {R} package for {Bayesian} spatial modelling from ecological survey data},
-  journal = {Methods in Ecology and Evolution},
-  volume = {10},
-  number = {},
-  pages = {760--766},
-  year = {2019},
-  month = {},
-  doi = {10.1111/2041-210X.13168},
-  publisher = {British Ecological Society}
-}
-
-@Online{CORDIS:2024,
-  author = {{CORDIS}},
-  title = {Global land ice, hydrology and ocean mass trends  |  GlobalMass},
-  journal = {},
-  volume = {},
-  number = {},
-  pages = {},
-  year = {2024},
-  month = {},
-  doi = {10.3030/694188},
-  url = {https://doi.org/10.3030/694188},
-  urldate = {2024-05-17}
-}
-
-@Misc{Lindgren:2024a,
-      author={{Lindgren}, F. and {Bachl}, F.E. and {Illian}, J. and {Suen}, M.H. and {Rue}, H. and {Seaton}, A.E.},
-      title={inlabru: software for fitting latent Gaussian models with non-linear predictors},
-      year={2024a},
-      eprint={2407.00791},
-      archivePrefix={arXiv},
-      primaryClass={stat.ME},
-      url={https://arxiv.org/abs/2407.00791}, 
-}
-
-@Misc{Lindgren:2024b,
-  author = {{Lindgren}, F. and {Bachl}, F.E. and {Borchers}, D.L. and {Simpson}, D. and {Scott-Howard}, L. and {Seaton}, A. and {Suen}, M.H. and {Roudier}, P. and {Meehan}, T. and {Peddinenikalva}, N. and {Perepolkin}, D.},
-  title = {inlabru},
-  year = {2024b},
-  month = {},
-  journal = {GitHub repository},
-  url = {https://inlabru-org.github.io/inlabru},
-  publisher = {GitHub}
-}
-
-@Article{Rue:2009,
-  author = {{Rue}, H. and {Martino}, S. and {Chopin}, N.},
-  title = {Approximate Bayesian inference for latent Gaussian models by using integrated nested Laplace approximations},
-  journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
-  volume = {71},
-  number = {2},
-  pages = {319-392},
-  year = {2009},
-  month = {},
-  doi = {10.1111/j.1467-9868.2008.00700.x},
-  url = {https://doi.org/10.1111/j.1467-9868.2008.00700.x},
-}
-
-@book{Vygotsky:1978,
-  author = {{Vygotsky}, L.S.},
-  title = {Mind in Society: Development of Higher Psychological Processes},
-  year = {1978},
-  isbn = {9780674576285},
-  url = {http://www.jstor.org/stable/j.ctvjf9vz4},
[…]
