### 2024-07-26 1c69acb
Addition of new reference (Lindgren et al., 2024 arXiv)
