### 2024-07-28 #7047 @editorialbot
[REVIEW]: 4DModeller: a spatio-temporal modelling package
**Submitting author:** @rwestaway (<a href="http://orcid.org/0000-0001-6102-1540">Richard Westaway</a>)
**Repository:** https://github.com/4DModeller/fdmr
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.1
