### 2024-07-28 #7047 @Nikoleta-v3
I also want to bring into this issue the discussion we were having with @rwestaway.

My comment:

> I see that you have some tests. Could you include some information on how to run the test suite? I really like your contribution guidelines; they are short and to the point. Maybe adding the test information there? or would be a bit much?

> Whatever you decide, it would just be useful to know how to run the tests locally since the tests run when people open pull requests.

The response of:

> @Nikoleta-v3 - To respond to your query, the testing suite is a relic of some old code so thank you for flagging that! At the time we had some ideas for testing meshes, and other topics, but as it stands we haven't really built it out or developed it further. However it is something planned for future work.

Thanks for clarifying. Having a test suite is not a requirement for publication with JOSS. It would be great to see some new tests being built.

Regarding the old code, I am not sure how you decided to handle this, but are you removing it? Currently it's a bit misleading.
