### 2024-08-01 #7047 @wcjochem
Hi @rwestaway, before I get into my review, can you please clarify Contribution and Authorship for this submission (see the 'General checks' for JOSS)? I see you are not the corresponding author on the paper and you have not made substantial commits to the repository. Can you please describe your contribution to the project and this submission? Have all authors agreed to submit this work to JOSS? Thanks
