### 2024-08-01 #7047 @rwestaway
Hi @wcjochem.

Thanks for agreeing to review. Please see below for author contributions based on the CRediT taxonomy – my personal contribution has been that I managed the project that produced the 4DModeller software, and I wrote the paper we submitted. All authors (cc-ed) are happy to submit this work to JOSS.

@John Mark ***@***.***> - Investigation, Software, Methodology, Writing – review & editing, Funding acquisition
@Gareth ***@***.***> – Software, Methodology
@Yin ***@***.***> - Investigation, Software, Methodology
@Anrijs ***@***.***> - Investigation, Software, Methodology
@Christopher ***@***.***> – Methodology, Supervision
@Richard ***@***.***> - Writing – original draft, Methodology, Project administration, Supervision
@Jonathan ***@***.***> – Conceptualization, Methodology, Supervision, Writing – review & editing, Funding acquisition

Please also note that I am on leave from today until 13th August. I suggest that @John Mark ***@***.***> and @Anrijs ***@***.***> are the best points of contact, if needed, during that time.

Many thanks
Richard

From: Chris Jochem ***@***.***>
Sent: Thursday, August 1, 2024 2:46 PM
Cc: Richard Westaway ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: 4DModeller: a spatio-temporal modelling package (Issue #7047)

Hi @rwestaway<https://github.com/rwestaway>, before I get into my review, can you please clarify Contribution and Authorship for this submission (see the 'General checks' for JOSS)? I see you are not the corresponding author on the paper and you have not made substantial commits to the repository. Can you please describe your contribution to the project and this submission? Have all authors agreed to submit this work to JOSS? Thanks
