### 2024-08-28 4DModeller/fdmr#324 @PieterjanRobbe
JOSS manuscript review
This issue entails my review comments for the JOSS manuscript submission at [openjournals/joss-reviews/issues/7047](https://github.com/openjournals/joss-reviews/issues/7047).

JOSS requires a "State of the field: Do the authors describe how this software compares to other commonly-used packages?". I couldn't explicitly find this in the paper. Does that mean that 4DModeller provides unique functionality that can't be found elsewhere?

Also, if I'm really nitpicking, I noticed that `fdmr` software package is written in typewriter font, but other software packages, e.g., `inlabru`, are written in normal font. Perhaps you could use typewriter font everywhere?

Thanks!
