### 2024-08-28 4DModeller/fdmr#325 @PieterjanRobbe
Notebooks depend on "INLA" package
This issue entails my review comments for the JOSS submission at [openjournals/joss-reviews/issues/7047](https://github.com/openjournals/joss-reviews/issues/7047).

It seems like running the [tutorials](https://4dmodeller.github.io/fdmr/articles/) requires the R package "INLA". I think it would be good to add a link to the installation instructions for that package somewhere, as it took me a while to understand that simply doing `install.packages("INLA")` doesn't work.

Thanks!
