### 2024-08-29 4DModeller/fdmr#326 @wcjochem
JOSS Review - fdmr::retrieve_tutorial_data should not modify the home directory
fdmr::retrieve_tutorial_data() should not modify my home directory and I should have an option to choose where data should be downloaded.

**Describe the solution you'd like**
I think there should be an argument to the function to specify the download location, and the default download location should be the temporary directory.

**Additional context**
If you have aspirations for this package to go to CRAN then it cannot write anywhere except for R's temporary directory by default.

openjournals/joss-reviews#7047
