### 2024-08-29 4DModeller/fdmr#327 @wcjochem
JOSS Review: error in model_viewer description
**Describe the bug**
The documentation for fdmr::model_viewer is incorrect and is a duplicate of fdmr::mesh_builder

**Expected behavior**
Updated description of the function to match its use.

**To Reproduce**
Check: https://github.com/4DModeller/fdmr/blob/main/R/shiny_modelviewer.R#L298

openjournals/joss-reviews#7047
