### 2024-09-04 4DModeller/fdmr#328 @wcjochem
JOSS Review: Functionality documentation
>Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I feel the package would benefit from more fully documented functions. While there is a large collection of vignettes, the individual functions have only a minimum template of documentation with argument names and a 1 sentence description. Given that all the functions in the package expect specific formats and types of inputs, I would expect that the type of arguments would be in the documentation. Further expanding the parts of the return values (i.e. more detail than just "list") would also help users.

openjournals/joss-reviews#7047
