### 2024-09-04 4DModeller/fdmr#329 @wcjochem
JOSS Review: State of the field
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

The manuscript currently does not provide a sufficient description, comparison, and references to the understand the state of the field. While the manuscript does mention INLA and inlabru (which this package works with), there is not a clear presentation of the features and other common packages for Bayesian spatial statistical modelling. Expanding the state of the field should also help with the framing to identify the clear statement of need and the gap that this project is filling.

openjournals/joss-reviews#7047
