### 2024-09-04 #7047 @wcjochem
Hi @Nikoleta-v3,

Thanks for the opportunity to review this software. This project, 4Dmodeller, provides three shiny apps to support researchers applying a particular type of spatial statistical model. It does this by wrapping up and using INLA and inlabru to support a limited set of Bayesian hierarchical model forms. The R package also provides common helper functions for parsing model output, visualising data, and some useful data cleaning/conversions.

The broader project associated with the software provides a large set of "vignettes" (long-form, documentation and worked examples) that are quite valuable for anyone learning to apply these types of models. However, I note that many of those vignettes are focused on general model fitting with inlabru and INLA software and less attention is sometimes paid to the fdmr software.

I've opened a few issues with comments on the repo. My main concern is that the documentation for the functions in the package would benefit from more detail. The functions expect the parameters to be in particular forms and, which this is covered in the vignettes, it's not in the function help pages.

Regarding the manuscript, it is quite short and I don't think it currently provides a sufficient level of description particularly on the state of the field and what this software contributes to it.

At this stage, I feel I've finished my review until the authors revise the work. Thanks!
