### 2024-09-04 #7047 @rwestaway
Thanks for the helpful comments @wcjochem! We will work through the issues you have opened. Regarding the manuscript length, we deliberately kept it as short as possible (it's currently ~800 words) in accordance with the JOSS submission guidelines [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), but we can certainly look to include further detail on the state of the field and what the software contributes.
