### 2024-09-04 #7047 @PieterjanRobbe
Hi @Nikoleta-v3, @rwestaway,

Same for me, I've started my review and opened a few issues. I'll continue my review as soon as these get resolved.

Perhaps it's also worth mentioning that I had a lot of issues installing the dependencies of 4DModeller on my M1 Mac.

Thanks!
