### 2024-09-21 73e75e1
Added the (default) option to save to the session's temporary directory
