### 2024-09-21 a7223ee
Improved the option so that the user can pass a character. Also added a checker if the character is a valid pass.
