### 2024-09-21 0ce34ce
Added a line to call for tutorial data retrieval - I guess testing has new temporary directory for each vignette rebuild.
