### 2024-09-22 3611546
Simplified so that we use the functions we already have
