+++ R/util_retrieve.R
@@ -11,7 +11,7 @@ retrieve_tutorial_data <- function(dataset, force_update = FALSE, save = FALSE) 
 
   if (base::isTRUE(save) | base::is.character(save)){
     if (base::is.character(save)) {
-      save_path <- check_path(save)
+      save_path <- clean_path(save, check_exists = TRUE)
       download_cache_folder <- fs::path(save_path, "fdmr", "download_cache")
     } else{
       download_cache_folder <- fs::path(fs::path_home(), "fdmr", "download_cache")
@@ -53,7 +53,7 @@ retrieve_tutorial_data <- function(dataset, force_update = FALSE, save = FALSE) 
 
   if (base::isTRUE(save) | base::is.character(save)){
     if (base::is.character(save)){
-      save_path <- check_path(save)
+      save_path <- clean_path(save, check_exists = TRUE)
       extract_path <- fs::path(save_path, "fdmr", "tutorial_data", dataset)
     } else{
       extract_path <- fs::path(fs::path_home(), "fdmr", "tutorial_data", dataset)
@@ -93,25 +93,4 @@ retrieve_tutorial_data <- function(dataset, force_update = FALSE, save = FALSE) 
   } else {
     stop("Invalid dataset, please see available datasets at https://github.com/4DModeller/fdmr_data")
   }
-}
-
-#' Checks if a path exists and returns an absolute path
-#'
-#' @param path Path to check
-#'
-#' @return fs::path Expanded absolute path
-#' @keywords internal
-#' 
-check_path <- function(path) {
-  if (is.null(path) || nchar(path) == 0) {
-    stop("Invalid path of zero length given.")
-  }
-  
-  fpath <- fs::path_abs(fs::path_expand(path))
-  
-  if (!fs::file_exists(fpath)) {
-    stop(paste("The path", fpath, "does not exist."))
-  }
-  
-  return(fpath)
 }
