### 2024-09-22 7d6fead
Updated plot_priors documentation
