### 2024-09-22 e672be9
Added more information to the docs
