### 2024-09-23 e951927
Updated plot_graph
