### 2024-09-23 efa4144
Another small update to plot_graph
