+++ man/mesh_builder.Rd
@@ -16,7 +16,7 @@ mesh_builder(
 )
 }
 \arguments{
-\item{spatial_data}{Spatial data}
+\item{spatial_data}{Spatial data (must be a data.frame, SpatialPointsDataFrame or SpatialPolygonsDataFrame)}
 
 \item{obs_data}{Measurement data}
 
@@ -33,7 +33,7 @@ mesh_builder(
 \item{x_coord}{Name of the longitude column in the spatial data}
 }
 \value{
-shiny::app
+shiny::app with meshbuilder functionality
 }
 \description{
 Mesh building shiny app. Creates and visualises a mesh from some spatial data.
