+++ man/model_viewer.Rd
@@ -12,16 +12,16 @@ model_viewer(
 )
 }
 \arguments{
-\item{model_output}{INLA model output}
+\item{model_output}{INLA/inlabru model output}
 
-\item{mesh}{INLA mesh}
+\item{mesh}{INLA/fmesher mesh used for the run}
 
-\item{measurement_data}{Measurement data}
+\item{measurement_data}{Measurement data used for the run (can be data.frame, SpatialPoints/SpatialPointsDataFrame, or simple feature collection)}
 
-\item{data_distribution}{Type of data, Poisson or Gaussian}
+\item{data_distribution}{Type of data, Poisson or Gaussian (character)}
 }
 \value{
-shiny::app
+shiny::app with model viewer functionality
 }
 \description{
 Mesh building shiny app. Creates and visualises a mesh from some spatial data.
