### 2024-09-24 #7047 @rwestaway
Hello. Yes, @mnky9800n is coordinating our responses to the issues raised, and some have already been addressed. I know he has been pre-occupied running a workshop this week, but I believe he intends to pick it up again from next week so we'll be able to provide a better update then.
