### 2024-10-12 4DModeller/fdmr#329 @mnky9800n
Hi, please see below (will be updated):

State of the field

There are number of packages within the R ecosystem that perform various spatio-temporal functions (e.g.,   [spTimer](https://cran.r-project.org/package=spTimer) MCMC modeling for space-time data, [spBayes](https://cran.r-project.org/package=spBayes) MCMC modeling of random effects for space-time data, and [bmstdr](https://cran.r-project.org/web/packages/bmstdr/index.html) a model comparison tool that uses MCMC, INLA, and other methods to fit to data and then compare between models). A critical expectation of each of these packages is a deep background in both bayesian modeling and understanding bayesian data. Users are expected to build their own fit equations, have a deep understanding of random effects, and are expected to be able to interpret predictions and uncertainties. However, none of these packages offer much if any scaffolding to do so. 4D-Modeller, attempts to bridge this gap with it's shiny apps.
