### 2024-10-12 e5ff9e1
added install INLA to the quickstart
