### 2024-10-12 #7047 @mnky9800n
Hi, i added some comments on the state of the field but i am unsure how much to write there, and how exhaustive this should be. an important component of 4d-modeller is that it is an onboarding tool into the bayesian spatio-temporal modeling space. getting from 0 to somewhere in most of the packages is an exponential learning curve and we were hoping to flatten that somewhat. so the state of the field is that there isn't necessarily anyone else thinking about it in this way, to our knowledge at least.
