+++ README.md
@@ -30,7 +30,7 @@ The easiest way to do this is to [install RStudio](https://posit.co/downloads/).
 Next start an R session and run
 
 ```R
-install.packages("INLA")
+install.packages("INLA", repos=c(getOption("repos"), INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE)
 install.packages("devtools")
 library(devtools)
 devtools::install_github("4DModeller/fdmr")
