### 2024-10-22 #7047 @Nikoleta-v3
Hello all 👋🏻 I’d like to get this review moving again 🏃🏻  From what I can see, a few comments have now been addressed.

@wcjochem @PieterjanRobbe, do you have any feedback on: https://github.com/4DModeller/fdmr/issues?

@mnky9800n @rwestaway, once you’ve finished your corrections, please write a comment here summarizing the changes and how you addressed each comment. As far as I can tell, most issues raised by the reviewers have been closed, but without clarification on the implementation or how they were resolved 😄
