### 2024-10-22 6c9d0b2
Update paper.md
Revised paper with additional 'state of field' text suggested by JA
