### 2024-10-22 70fe089
Update paper.bib
Updated references
