### 2024-10-23 #7047 @mnky9800n
hi, i merged all the pull requests so at this point i think it is okay to
move on from your end.

On Tue, Oct 22, 2024 at 5:56 PM Nikoleta Glynatsi ***@***.***>
wrote:

> Hello all 👋🏻 I’d like to get this review moving again 🏃🏻 From what I
> can see, a few comments have now been addressed.
>
> @wcjochem <https://github.com/wcjochem> @PieterjanRobbe
> <https://github.com/PieterjanRobbe>, do you have any feedback on:
> https://github.com/4DModeller/fdmr/issues?
>
> @mnky9800n <https://github.com/mnky9800n> @rwestaway
> <https://github.com/rwestaway>, once you’ve finished your corrections,
> please write a comment here summarizing the changes and how you addressed
> each comment. As far as I can tell, most issues raised by the reviewers
> have been closed, but without clarification on the implementation or how
> they were resolved 😄
>
