### 2024-11-15 4DModeller/fdmr#337 @PieterjanRobbe
Timeout error when downloading INLA dependency
Hi there,

I'm having some issues trying to install the dependencies for this package. It seems like the command below runs into a timeout error. I've tried to install the package several times over the past few days, but the same error keeps popping up. This is not related to a slow internet connection (other downloads are fine). Any idea what I'm doing wrong?

Thanks!

```
> install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/stable"), dep = TRUE)
Warning in install.packages :
  dependencies ‘Rgraphviz’, ‘graph’, ‘INLAspacetime’ are not available

  There is a binary version available but the source version is later:
       binary   source needs_compilation
INLA 23.04.24 24.06.27             FALSE

installing the source package ‘INLA’

trying URL 'https://inla.r-inla-download.org/R/stable/src/contrib/INLA_24.06.27.tar.gz'
Content type 'application/x-gzip' length 62225912 bytes (59.3 MB)
=
downloaded 1.2 MB

Warning in install.packages :
  downloaded length 1263866 != reported length 62225912
Warning in install.packages :
  URL 'https://inla.r-inla-download.org/R/stable/src/contrib/INLA_24.06.27.tar.gz': Timeout of 60 seconds was reached
Error in download.file(url, destfile, method, mode = "wb", ...) :
  download from 'https://inla.r-inla-download.org/R/stable/src/contrib/INLA_24.06.27.tar.gz' failed
Warning in install.packages :
  download of package ‘INLA’ failed
```

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/7047)
