### 2024-11-15 4DModeller/fdmr#337 @finnlindgren
You need to install the prebuilt binaries, and not the source version. Also, make sure you have R 4.4 and use
```
install.packages("INLA", repos = c(getOption("repos"), INLA = "https://inla.r-inla-download.org/R/testing"), dep = TRUE)
```
i.e. testing, not stable.

In addition, your internet connection seems slow, so R times out the file download. use e.g. `options(timeout = 3600)` to set a longer timeout than 60 seconds.
