### 2024-11-15 4DModeller/fdmr#337 @finnlindgren
In some cases on some platforms, R may need to be told to ignore the newer source version, and force it to install a binary version. The https://www.r-inla.org/download-install mentions a related issue for macOS & homebrew.
