### 2024-11-16 4DModeller/fdmr#337 @PieterjanRobbe
Hi @finnlindgren, thanks, changing from stable to testing did the trick!

I was following the instructions from the README file. Perhaps it's a good idea to update the README to reflect this change?
