### 2024-11-16 4DModeller/fdmr#338 @PieterjanRobbe
Shiny app in mesh building tutorial is stuck
Hello,

I'm trying to go through the mesh building tutorial [here](https://4dmodeller.github.io/fdmr/articles/meshbuilder.html). I'm trying to change some of the initial parameters. This seems to work well when I open the app like this:

```
fdmr::mesh_builder(spatial_data = sp_data, max_edge = max_edge_fin, offset = offset, cutoff = cutoff)
```

However, when I open the app with

```
fdmr::mesh_builder(spatial_data = sp_data)
```

and then click `Enable customisation` and manually change one of the parameters, the app just appears to hang. I'm wondering if I'm doing something wrong?

Also, I think there's a broken sentence above [this](https://4dmodeller.github.io/fdmr/articles/meshbuilder.html#customise-initial-parameters) paragraph:

> This object contains all the contains all the values we We can build a mesh

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/7047#issuecomment-2458925362)
