### 2024-11-16 4DModeller/fdmr#339 @PieterjanRobbe
Model builder tutorial
Hi there,

I'm working my way through the [Model builder](https://4dmodeller.github.io/fdmr/articles/modelbuilder.html) tutorial. After launching the Shiny app in [this](https://4dmodeller.github.io/fdmr/articles/modelbuilder.html) section, I follow the instructions in the tutorial. I go to the `Features` tab and toggle the `cases` checkmark. Then, under the `Model` tab, I click `Run`, and I get this error:

```
 *** inla.core.safe:  The inla program failed, but will rerun in case better initial values may help. try=1/1
Loading required package: INLA
Loading required package: Matrix
This is INLA_24.11.07-4 built 2024-11-07 10:56:07 UTC.
 - See www.r-inla.org/contact-us for how to get help.
 - List available models/likelihoods/etc with inla.list.models()
 - Use inla.doc(<NAME>) to access documentation
Warning in iinla(model = info[["model"]], lhoods = info[["lhoods"]], options = info[["options"]]) :
  iinla: Problem in inla: Error in inla.core.safe(formula = formula, family = family, contrasts = contrasts,  :
  The response for family[1] is not of type 'numeric|list|matrix'; don't know what to do.
The inla program failed and the maximum number of tries has been reached.
iinla: Problem in inla: 1: workRSOCK()
2: workLoop(makeSOCKmaster(master, port, setup_timeout, timeout,
       useXDR, setup_strategy))
3: workCommand(master)
4: tryCatch({
       msg <- recvData(master)
       if (msg$type == "DONE") {
           closeNode(master)
           FALSE
       }
       else if (msg$type == "EXEC") {
           success <- TRUE
           handler <- function(e) {
               success <<- FALSE
               structure(conditionMessage(e), class = c("snow-try-error",
                   "try-error"))
           }
           t1 <- proc.time()
           value <- tryCatch(do.call(msg$data$fun, msg$data$args,
               quote = TRUE), error = handler)
           t2 <- proc.time()
      [... truncated]
iinla: Giving up and returning last successfully obtained result for diagnostic purposes.
Warning: Error in create_prediction_field: var_b must be provided for predicted_mean_fields plot
  82: stop
  81: create_prediction_field
  80: <reactive>
  64: prediction_field
  63: <reactive>
  47: z_values
  46: ::
shiny
observe
  45: <observer>
   2: shiny::runApp
   1: fdmr::model_builder
```

Am I missing something?

[breadcrumb](https://github.com/openjournals/joss-reviews/issues/7047#issuecomment-2458925362)
