### 2024-11-16 2b45c95
Update README.md
Changed it from the "stable" to "testing". It shouldn't cause issues as the R-CMD-check already has the testing version as a requirement.
