+++ README.md
@@ -30,13 +30,15 @@ The easiest way to do this is to [install RStudio](https://posit.co/downloads/).
 Next start an R session and run
 
 ```R
-install.packages("INLA", repos=c(getOption("repos"), INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE)
+install.packages("INLA", repos=c(getOption("repos"), INLA="https://inla.r-inla-download.org/R/testing"), dep=TRUE)
 install.packages("devtools")
 library(devtools)
 devtools::install_github("4DModeller/fdmr")
 ```
 
 You should now have `fdmr` and all its dependencies installed and you can continue on [one of our tutorials](https://4dmodeller.github.io/fdmr/articles/).
+
+If your installation of `INLA` fails, please consult the [R-INLA website](https://www.r-inla.org/download-install) for platform-specific instructions.
 
 ## Installation
 
