### 2024-11-19 4DModeller/fdmr#339 @PieterjanRobbe
Hi @mnky9800n. No, I literally copy-pasted from the tutorial. Here's the model specification:
