### 2024-12-18 c5f64d3
Updated the docs. Passing a "data.frame" to measurement_data in "model_builder" would fail with the latest changes.
