### 2024-12-18 fb21bec
Removing the non-numeric or non-logical variables from the model features options
