### 2024-12-18 c8af62d
Revert "Removing the non-numeric or non-logical variables from the model features options"
This reverts commit 595a86b96eae95a9903a20455edaa0d25a6fa2cd.
