### 2024-12-18 c22ce63
Revert "Updated the docs. Passing a "data.frame" to measurement_data in "model_builder" would fail with the latest changes."
This reverts commit c5f64d3e9cf3bb86181647931b7347e3e6ad6046.
