### 2024-12-18 4DModeller/fdmr#338 @aabelean
Hi @PieterjanRobbe, sorry for the delay getting back to you!
I didn't find an issue with the ShinyApp itself, but the default values for `max.edge` are way too small for the Tibetan lake case. I think it hangs because the mesh is large. I couldn't even run it on my laptop with the default or similar parameters outside of the app. To clarify further, the defaults we use in the app aren't the same as those provided by `fmesher` and used in the initial mesh (that shows up when you open the app).

Regarding the second issue:
> Also, I think there's a broken sentence above [this](https://4dmodeller.github.io/fdmr/articles/meshbuilder.html#customise-initial-parameters) paragraph:

Thanks for noticing this, I also corrected this in the pull request.
