### 2024-12-18 4DModeller/fdmr#339 @aabelean
Hi @PieterjanRobbe,

I think the issue is that the variable "UtlaName" (and also MSOA11CD) is a string variable. `INLA` (and `inlabru`) does not accept them for this purpose. I suggest we no longer offer non-numeric/non-logical variables to be selected as model/exposure variables or features.
