### 2024-12-20 4DModeller/fdmr#339 @mnky9800n
This is true and should be taken care of by adding a bit of logic to the app
