### 2025-01-04 #7047 @Nikoleta-v3
Hello everyone, and Happy New Year🎆

I’ve noticed there are still open issues, small bugs being fixed, and ongoing discussions. Thank you all for your hard work!
