### 2025-01-07 #7047 @rwestaway
Hello @Nikoleta-v3 and Happy New Year to you too! We looked at the open issues just before Christmas and they have now been resolved and closed.
