### 2025-01-08 #7047 @wcjochem
Hi @Nikoleta-v3, I can see in the commits that most of the issues I raised are getting addressed. There is still an open issue, and I would like to see the paper updated (a new version didn't seem to build above) and some updates are still on different branches. After that I should be able to complete my checklist. Thanks.
