### 2025-01-09 #7047 @Nikoleta-v3
Thank you for your comment.

@rwestaway @mnky9800n, could you please rebase your current branch `JOSS` to include the new additions from the `devel` branch, as these changes are necessary for the documentation and functionality of the package?

In my understanding, both reviewers have raised questions about the state of the field, and the manuscript currently does not provide a sufficient description, comparison, or references to understand the state of the field. To address these comments, the authors included a paragraph on other packages in this commit: https://github.com/4DModeller/fdmr/pull/335/commits/6542645d6ef5d4f356e52c0d95b4966917b78689.

These were the only changes made to the paper, and they appear in the newly compiled paper here: https://github.com/openjournals/joss-reviews/issues/7047#issuecomment-2578660669. However, I might be missing something. Let me investigate further.

One of the reviews states the following:

> Also, if I'm really nitpicking, I noticed that the `fdmr` software package is written in typewriter font, but other software packages, e.g., `inlabru`, are written in normal font. Perhaps you could use typewriter font everywhere?

However, this issue was not addressed in the revision.

Please, when you make changes or address comments, either properly reply to the issues or write a concise summary here so we can all stay on the same page.
