### 2025-01-10 #7047 @rwestaway
Thanks for the messsage. We will make the two changes you mention (rebase current branch and add typewriter font) and then confirm and provide details here.

You are correct to identify that we tried to address the questions about state of the field by adding text into the manuscript: "_There are number of packages within the R ecosystem that perform various spatio-temporal functions (e.g., spTimer MCMC modeling for space-time data, spBayes MCMC modeling of random effects for space-time data, and bmstdr a model comparison tool that uses MCMC, INLA, and other methods to fit to data and then compare between models). A critical expectation of each of these packages is a deep background in both bayesian modeling and understanding bayesian data. Users are expected to build their own fit equations, have a deep understanding of random effects, and are expected to be able to interpret predictions and uncertainties. However, none of these packages offer much if any scaffolding to do so._".

Please advise if this is sufficient of if the reviewers would like further detail.
