### 2025-01-13 #7047 @mnky9800n
also see the follow up commit with the proper `paper.md` file from @rwestaway
