### 2025-01-13 #7047 @rwestaway
So to summarise:
- @mnky9800n has rebased JOSS (assuming this has worked ok?)
- We have updated the manuscript with (i) additional text about state of the field and (ii) to apply `typewriter` font across all R package names
