### 2025-01-20 #7047 @Nikoleta-v3
Hello both,

I believe the rebase didn’t work, or perhaps you didn’t push the changes?

**Why do I think it hasn’t worked?**
- On the `JOSS` branch, it says “[92 commits behind](https://github.com/4DModeller/fdmr/compare/JOSS...devel) devel.”
- Additionally, the [R/shiny_meshbuilder.R](https://github.com/openjournals/joss-reviews/issues/7047#diff-1e62ab8249d1c140aa8d030ecd8c74bfa8c3e940a37f656db1d2e987feff9df1) file has been updated during the review process. However, when you view the file on the `JOSS` branch (here: https://github.com/4DModeller/fdmr/blob/JOSS/R/shiny_meshbuilder.R), the changes don’t appear.

Regarding the reviewers’ comments on the paper, could you please make sure to properly reply to the issues? Specifically, issues `324` and `329`. This will allow the reviewers to see your responses and close the issues if they are satisfied with the changes.
