### 2025-01-21 4DModeller/fdmr#329 @richardwestaway
We have updated the manuscript with additional text about state of the field:
_"There are number of packages within the R ecosystem that perform various spatio-temporal functions (e.g., spTimer MCMC modeling for space-time data, spBayes MCMC modeling of random effects for space-time data, and bmstdr a model comparison tool that uses MCMC, INLA, and other methods to fit to data and then compare between models). A critical expectation of each of these packages is a deep background in both bayesian modeling and understanding bayesian data. Users are expected to build their own fit equations, have a deep understanding of random effects, and are expected to be able to interpret predictions and uncertainties. However, none of these packages offer much if any scaffolding to do so."_
