### 2025-01-21 #7047 @mnky9800n
Hi, I think you are correct, i never resolved the conflicts between JOSS and devel (because of the changelog.md so it really wasn't anything big). I have done that now and hopefully this fixes this issue. I believe @rwestaway replied to those issues now.
