### 2025-02-05 #7047 @rwestaway
Hi. Any update on this paper? We're not sure whether you're waiting on us to do more? Or whether we should be expecting further reviewer comments? Or whether all issues are now closed meaning it can now be accepted? Thanks!
