### 2025-02-07 #7047 @Nikoleta-v3
Hello @rwestaway 👋🏻 thanks for your question. The issues are not yet closed, but thank you for commenting on them. We’re just waiting for the reviewers to give their final okay, then we can move ahead with the submission.

@wcjochem @PieterjanRobbe, did you have a chance to look at the latest reply from the authors?
