### 2025-02-13 #7047 @Nikoleta-v3
👋🏻 Thank you for your patience, @rwestaway! I opened a pull request that fixes the formatting of some references here: https://github.com/4DModeller/fdmr/pull/345.

If you agree, please go ahead and merge.

At this point could you please:

- [ ] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release  v0.2.0
