+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   calving_tutorial: calving_tutorial.html
@@ -19,7 +19,7 @@ articles:
   simulation_gaussian_data: simulation_gaussian_data.html
   simulation_poisson_data: simulation_poisson_data.html
   tutorial_geophysics: tutorial_geophysics.html
-last_built: 2024-07-11T11:34Z
+last_built: 2025-02-17T13:21Z
 urls:
   reference: https://4dmodeller.github.io/fdmr/reference
   article: https://4dmodeller.github.io/fdmr/articles
