+++ pkgdown.yml
@@ -19,7 +19,7 @@ articles:
   simulation_gaussian_data: simulation_gaussian_data.html
   simulation_poisson_data: simulation_poisson_data.html
   tutorial_geophysics: tutorial_geophysics.html
-last_built: 2025-02-17T13:21Z
+last_built: 2025-02-17T13:32Z
 urls:
   reference: https://4dmodeller.github.io/fdmr/reference
   article: https://4dmodeller.github.io/fdmr/articles
