### 2025-02-18 #7047 @rwestaway
[] If you agree, please go ahead and merge.

@mnky9800n has done this

_[x] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release v0.2.0_

@mnky9800n has done this - it's now v0.2.1

_[x] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)_

Done - in Zenodo

_[x] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID._

Done - hopefully all metadata is now correct but please let us know if you spot any discrepancies

_[x] Please list the DOI of the archived version here._

Zenodo DOI of v0.2.1 is https://doi.org/10.5281/zenodo.14883337
