### 2025-02-19 #7047 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
