### 2025-02-19 8068345
minor changes in cases in paper.bib
