+++ man/paper.bib
@@ -1,6 +1,6 @@
 @article{Aiken:2018,
   author={{Aiken}, J.M. and {Aiken}, C. and {Cotton}, F.},
-  title = "{A Python Library for Teaching Computation to Seismology Students}",
+  title = {A {P}ython Library for Teaching Computation to Seismology Students},
   journal = {Seismological Research Letters},
   volume = {89},
   number = {3},
@@ -42,7 +42,7 @@
 
 @Article{Lindgren:2024a,
       author={{Lindgren}, F. and {Bachl}, F.E. and {Illian}, J. and {Suen} and {Rue}, H. and {Seaton}, A.E.},
-      title={{inlabru}: software for fitting latent Gaussian models with non-linear predictors}, 
+      title={{inlabru}: software for fitting latent {G}aussian models with non-linear predictors}, 
       year={2024},
       eprint={2407.00791},
       journal = {arXiv},
@@ -62,7 +62,7 @@
 
 @Article{Rue:2009,
   author = {{Rue}, H. and {Martino}, S. and {Chopin}, N.},
-  title = {Approximate Bayesian inference for latent Gaussian models by using integrated nested Laplace approximations},
+  title = {Approximate {B}ayesian inference for latent {G}aussian models by using integrated nested {L}aplace approximations},
   journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
   volume = {71},
   number = {2},
@@ -99,7 +99,7 @@
 
 @Article{Yin:2023,
   author = {{Yin}, X. and {Aiken}, J.M. and {Bamber}, J.L.},
-  title = {{4DModeller (fdmr): A Comprehensive R Package for Spatio-Temporal Modelling}},
+  title = {{4DModeller (fdmr)}: A Comprehensive {R} Package for Spatio-Temporal Modelling},
   journal = {Spatial Data Science Symposium 2023 Short Paper Proceedings},
   volume = {},
   number = {},
