### 2025-02-19 #7047 @rwestaway
@danielskatz - I have noticed a few remaining very minor issues:

Summary, para 1, lines 3 and 5 - As per previous reviewer comments, "inlabru" should be in typewriter font (i.e. `inlabru` in the md file). I must have missed those before....
Summary, para 1, line 3 - There is a missing comma. It should read "...R codes for fast, efficient Bayesian inference..."
Statement of Need, para 2, line 4 - There is a missing "a" . It should read "There are a number of packages..."

Is it best that all changes at this stage go through you?
