### 2025-02-19 #7047 @danielskatz
You can make those changes.
Also, in the Statement of Need, para 3, please add a "," after the "e.g." in the parentheses with Zeigler.
