### 2025-02-19 75b830e
Final minor changes in paper.md
