+++ man/paper.md
@@ -44,7 +44,7 @@ bibliography: paper.bib
 
 # Summary
 
-4DModeller (`fdmr`) is a spatio-temporal modelling package capable of solving a wide range of large-scale space-time (i.e., four-dimensional) problems [@Yin:2023]. It is built around the inlabru framework, which is a suite of R codes for fast efficient Bayesian inference [@Yuan:2017; @Bachl:2019; @Lindgren:2024a]. The `fdmr` package expands the inlabru framework to include specific applications of latent variable modelling for 4-D geophysical problems (e.g., ocean heat content, the Earth’s magnetic field, and global sea-level rise). `fdmr` also includes shiny apps that provide tools for data visualization, finite element mesh building, and Bayesian hierarchical modelling based on an R package for Bayesian inference, `inlabru`, along with model evaluation and assessment. These shiny apps are designed to make the complex INLA framework [@Rue:2009] and associated concepts accessible to a wider scientific community, including users who have little to no previous experience using R. The tools are designed with new users in mind by leveraging their expertise with their data sets while minimizing the need to develop extensive code in R [@Aiken:2018; @Vygotsky:1978]. They allow users to interact with their data first using the intuitive knowledge of the modelling process (input data, create mesh, calculate statistical model), then auto-generating code that the users can build on.
+4DModeller (`fdmr`) is a spatio-temporal modelling package capable of solving a wide range of large-scale space-time (i.e., four-dimensional) problems [@Yin:2023]. It is built around the `inlabru` framework, which is a suite of R codes for fast, efficient Bayesian inference [@Yuan:2017; @Bachl:2019; @Lindgren:2024a]. The `fdmr` package expands the `inlabru` framework to include specific applications of latent variable modelling for 4-D geophysical problems (e.g., ocean heat content, the Earth’s magnetic field, and global sea-level rise). `fdmr` also includes shiny apps that provide tools for data visualization, finite element mesh building, and Bayesian hierarchical modelling based on an R package for Bayesian inference, `inlabru`, along with model evaluation and assessment. These shiny apps are designed to make the complex INLA framework [@Rue:2009] and associated concepts accessible to a wider scientific community, including users who have little to no previous experience using R. The tools are designed with new users in mind by leveraging their expertise with their data sets while minimizing the need to develop extensive code in R [@Aiken:2018; @Vygotsky:1978]. They allow users to interact with their data first using the intuitive knowledge of the modelling process (input data, create mesh, calculate statistical model), then auto-generating code that the users can build on.
 
 This is extended through the Tutorial Driven Software Development practice [@Woods:2022]. This approach is designed to integrate subject matter experts into the code development cycle. It involves the identification of representative and instructive use cases, followed by tutorials that describe how the package could be used to solve them, and then finally code written and tested so that it behaves as described in the tutorials [@Woods:2022]. `fdmr` users have access to a set of domain-specific tutorials as vignettes in R Markdown notebooks; the tutorials are being added to as the user community grows.
 
@@ -56,9 +56,9 @@ Future `fdmr` package development efforts will focus on expanding its capabiliti
 
 The objective of the `fdmr` package is to provide researchers and practitioners with a straightforward and efficient solution for handling and analyzing different types of spatio-temporal data, enabling a comprehensive analysis of the underlying patterns, processes and trends.
 
[…]
