### 2025-02-19 #7047 @danielskatz
@rwestaway - please again check the proof that will be generated and let me know if it's now ok
